+++ src/store/defaultState.js
@@ -1,6 +1,6 @@
 const state = () => {
 	return {
-		webappVersion: '2.5.1',
+		webappVersion: '2.6.0',
 		userInfo: {
 			email: '',
 			firstName: '',